+++ services/web/server/src/simcore_service_webserver/invitations/invitations_core.py
@@ -12,7 +12,7 @@ from pydantic.errors import PydanticErrorMixin
 from servicelib.error_codes import create_error_code
 from simcore_postgres_database.models.users import users
 
-from .db import get_database_engine
+from ..db import get_database_engine
 from .invitations_client import (
     InvitationContent,
     InvitationsServiceApi,
