+++ services/web/server/tests/unit/with_dbs/03/invitations/conftest.py
@@ -14,8 +14,8 @@ from aiohttp import web
 from aiohttp.test_utils import TestClient
 from models_library.utils.fastapi_encoders import jsonable_encoder
 from pytest_simcore.aioresponses_mocker import AioResponsesMock
-from simcore_service_webserver.invitations_client import InvitationContent
-from simcore_service_webserver.invitations_settings import (
+from simcore_service_webserver.invitations.invitations_client import InvitationContent
+from simcore_service_webserver.invitations.invitations_settings import (
     InvitationsSettings,
     get_plugin_settings,
 )
